All Democracy Scores Dataset

A dataset containing nearly all extant democracy indexes. See the package \linkQuickUDS for further documentation.

democracy

Format

An object of class tbl_df (inherits from tbl, data.frame) with 24175 rows and 62 columns.

Variable descriptions

country_name

Standardized country name. This is the same across all datasets in this package, so you can always join them by country_name and year. Character with 225 distinct values. Most common: Afghanistan (269), Nepal (249), Oman (275). NAs = 0.

GWn

Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999). Numeric. Max = 990, min = 2, distinct = 225, mean = 440.49, sd = 255.909, NAs = 0.

GWc

Gleditsch-Ward alphabetic country code. See Gleditsch and Ward (1999). Character with 225 distinct values. Most common: AFG (269), NEP (249), OMA (275). NAs = 0.

cown

Correlates of War numeric country code. Differs from GWn for a few country-years. See Gleditsch and Ward (1999). Numeric. Max = 990, min = 2, distinct = 224, mean = 440.394, sd = 255.672, NAs = 0.

polity_ccode

Country code in Polity datasets. Differs from GWn for a few country-years. See Gleditsch and Ward (1999). Numeric. Max = 990, min = 2, distinct = 232, mean = 440.374, sd = 255.668, NAs = 0.

year

Year. Numeric. Max = 2015, min = 1741, distinct = 275, mean = 1935.172, sd = 57.84, NAs = 0.

GW_startdate

Date at which the state entered the system of states according to Gleditsch and Ward, or NA if it has never been a member. Date. Max = 2011-07-09, min = 1816-01-01, distinct = 171, NAs = 0.

GW_enddate

Date at which the state ceased to be a member of the system of states according to Gleditsch and Ward, or NA if it still exists. Date. Max = 2006-06-04, min = 1830-07-05, distinct = 30, NAs = 21097.

region

Region. Character with 23 distinct values. Most common: Eastern Africa (1990), South America (2136), Western Europe (2111). NAs = 0.

continent

Continent. Character with 5 distinct values. Most common: Africa (6269), Asia (5559), Europe (6384). NAs = 0.

microstate

Indicator of whether the state is a microstate, according to Gleditsch's list of microstates. Logical. TRUE = 1117, FALSE = 23058, NAs = 0.

lat

Latitude. Numeric. Max = 64.963, min = -40.901, distinct = 220, mean = 19.267, sd = 25.209, NAs = 0.

lon

Longitude. Numeric. Max = 178.68, min = -175.198, distinct = 220, mean = 14.893, sd = 66.221, NAs = 0.

in_system

Whether the country-year is in the Gleditsch-Ward system of states. See Gleditsch and Ward (1999). Logical. TRUE = 18253, FALSE = 5922, NAs = 0.

in_cow

Whether the country-year is in the Correlates of War system of states. Logical. TRUE = 16914, FALSE = 7261, NAs = 0.

arat_pmm

Democracy score Arat (1991). Taken from Pemstein, Meserve, and Melton (2013) replication data. Numeric. Max = 109, min = 29, distinct = 78, mean = 73.2, sd = 18.915, NAs = 20302.

blm

Trichotomous measure of regime type from Bowman, Lehoucq, and Mahoney (2005). 0 = authoritarian, 0.5 = semidemocratic, 1 = democratic. Available only for five Latin American countries (Costa Rica, El Salvador, Guatemala, Honduras and Nicaragua) from 1900 to 2000. Numeric. Max = 1, min = 0, distinct = 4, mean = 0.252, sd = 0.357, NAs = 23670.

bmr_democracy

Dichotomous measure of regime type from Boix, Miller, and Rosato (2012). Numeric. Max = 1, min = 0, distinct = 3, mean = 0.319, sd = 0.466, NAs = 7189.

bmr_democracy_omitteddata

Dichotomous measure of regime type from Boix, Miller, and Rosato (2012). 1 = democracy. This is the same measure as bmr_democracy, except it records an NA for countries occupied during an international war (e.g., the Netherlands 1940-44) or experiencing state collapse during a civil war (e.g., Lebanon 1976-89). The democracy variable instead fills in these years as continuations of the same regime type. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.321, sd = 0.467, NAs = 7414.

bnr

Dichotomous indicator of democracy from the Bernhard, Nordstrom & Reenock (2001). Event History Coding of Democratic Breakdowns. 0 = non-democracy, 1 - democracy. This indicator has been put in country-year format, extending to 1913, with the help of the Correlates of War panel of independent states; independent countries (not microstates) in this panel that were not included in the original dataset are assumed to be non-democratic for the period.? Numeric. Max = 1, min = 0, distinct = 3, mean = 0.347, sd = 0.476, NAs = 13823.

bollen_pmm

0-100 index of democracy from Bollen (2001). Taken from Pemstein, Meserve, and Melton (2013). Numeric. Max = 100, min = 0, distinct = 349, mean = 55.457, sd = 33.696, NAs = 23665.

doorenspleet

Dichotomous index of democracy from Doorenspleet (2000). 1 = authoritarian, 2 = democracy. Omits periods of interruption. Numeric. Max = 2, min = 1, distinct = 3, mean = 1.179, sd = 0.383, NAs = 11166.

eiu

0-1 index of democracy from the updated version of the Economist Intelligence Unit. 2012. Democracy Index 2012: Democracy at a Standstill. 0 = least democratic, 1 = most democratic. Taken from http://www.govindicators.org. Numeric. Max = 0.965, min = 0, distinct = 789, mean = 0.467, sd = 0.243, NAs = 21774.

freedomhouse

Average civil liberties + political rights score (reversed so higher values are more democratic) from Freedom House (2015). Goes from 1 (least democratic) to 7 (most democratic). In this version, the index does not include a value for 1981. This is based on the latest Freedom House data going all the way to 2015. Numeric. Max = 7, min = 1, distinct = 14, mean = 4.259, sd = 2.057, NAs = 16496.

freedomhouse_electoral

An indicator of whether a country is an "electoral democracy" in Freedom House's estimation (1 = yes, 0 - no). Original data available at http://www.freedomhouse.org. Goes from 1 (least democratic) to 7 (most democratic). Available only from 1989. This is based on the latest Freedom House data going all the way to 2015. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.599, sd = 0.49, NAs = 19103.

gwf

Dichotmous democracy/autocracy indicator from Geddes, Wright, and Frantz (2014). 0 = autocracy, 1 = democracy. Extended beyond 1945 using Geddes, Wright, and Frantz's case variable, which encodes information about the first year of the regime. Numeric. Max = 2, min = 1, distinct = 3, mean = 1.429, sd = 0.495, NAs = 14780.

hadenius_pmm

0-10 index of democracy from Hadenius 1992. Taken from Pemstein, Meserve, and Melton 2013. Higher values are more democratic. Numeric. Max = 10, min = 0, distinct = 52, mean = 4.509, sd = 3.56, NAs = 24046.

kailitz_binary

Dichotomous democracy indicator from Kailitz 2013. 1 = autocracy (all types including electoral autocracy), 2 = liberal democracy. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.412, sd = 0.492, NAs = 14569.

kailitz_tri

Trichotomous democracy indicator from Kailitz 2013. 1 = autocracy (all types except electoral autocracy), 2 = electoral autocracy, 3 = liberal democracy. Numeric. Max = 2, min = 0, distinct = 4, mean = 0.989, sd = 0.914, NAs = 14569.

lied

0-6 Lexical Index of Electoral Democracy from Skaaning, Gerring, and Bartusevicius. 2015. This is from V3 of the dataset, updated to 2015. 0 = nonelectoral, 1 = one- or no- party elections, 2 = limited competition multiparty elections for legislature only, 3 = Limited competition multiparty elections for both executive and legislature, 4 = Competitive elections for executive and legislative, limited suffrage, 5 = Male democracy, 6 = Electoral democracy. Numeric. Max = 6, min = 0, distinct = 8, mean = 2.783, sd = 2.346, NAs = 5646.

magaloni_democ_binary

Dichotomous democracy indicator from Magaloni, Chu, and Min (2013). 0 = autocracy (all types including multiparty autocracy), 1 = democracy. Extended beyond 1950 using the duration_nr variable of the original dataset, which encodes information about the first year of each regime. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.424, sd = 0.494, NAs = 13887.

magaloni_regime_tri

Trichotomous democracy indicator from Magaloni, Chu, and Min. 2013. 1 = autocracy (all types except multipary autocracy), 2 = multiparty autocracy, 3 = democracy. Extended beyond 1950 using the duration_nr variable of the original dataset, which encodes information about the first year of each regime. Numeric. Max = 3, min = 1, distinct = 4, mean = 2.011, sd = 0.915, NAs = 13887.

mainwaring

Trichotomous democracy indicator from Mainwaring, Brinks, and Perez Linan (2008). -1 = non-democracy, 0 = hybrid, 1 = democracy. Numeric. Max = 1, min = -1, distinct = 4, mean = -0.325, sd = 0.838, NAs = 22019.

munck_pmm

0-1 index of democracy from Munck 2009. Taken from Pemstein, Meserve, and Melton 2013. Only available for 342 country-years. Higher values are more democratic. Numeric. Max = 1, min = 0, distinct = 22, mean = 0.838, sd = 0.259, NAs = 23833.

pacl

Dichotomous measure of democracy from Cheibub, Gandhi, and Vreeland 2010. 1= democracy, 0 = non-democracy. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.438, sd = 0.496, NAs = 15060.

PEPS1i

Participation-Enhanced Polity Score 1, polity score adjusted using IDEA Votes/Voting age population. From Moon et al 2006. Numeric. Max = 10, min = -10, distinct = 727, mean = -3.199, sd = 6.574, NAs = 15012.

PEPS1q

Participation-Enhanced Polity Score 1, polity score adjusted using (mostly)IDEA votes/Voting age population, with participation coded zero for noncompetitive elections. From Moon et al 2006. Numeric. Max = 10, min = -10, distinct = 728, mean = -3.176, sd = 6.486, NAs = 14736.

PEPS1v

Participation-Enhanced Polity Score 1, polity score adjusted using Vanhanen votes/two-thirds of Vanhanen population. From Moon et al. (2006). Numeric. Max = 10, min = -10, distinct = 1867, mean = -2.869, sd = 5.661, NAs = 10086.

PEPS2i

Participation-Enhanced Polity Score 2, polity score adjusted using IDEA Votes/Voting age population. From Moon et al. 2006. Numeric. Max = 10, min = -9.59, distinct = 844, mean = 3.716, sd = 4.39, NAs = 20050.

PEPS2q

Participation-Enhanced Polity Score 2, polity score adjusted using (mostly)IDEA votes/Voting age population, with participation coded zero for noncompetitive elections. From Moon et al. 2006. Numeric. Max = 10, min = -10, distinct = 857, mean = -1.126, sd = 6.908, NAs = 17285.

PEPS2v

Participation-Enhanced Polity Score 2, polity score adjusted using Vanhanen votes/two-thirds of Vanhanen population. From Moon et al. 2006. Numeric. Max = 10, min = -10, distinct = 2415, mean = -2.427, sd = 5.777, NAs = 10934.

pitf

A five category indicator of democracy described in Goldstone et al 2010. Can be:

0-Full autocracy (exrec < 7, parcomp !=0 and parcomp < 3)

1-Partial autocracy (exrec < 7, parcomp = 0 or parcomp > 2)

2-Partial democracy with factionalism (exrec > 6, parcomp = 3)

3-Partial democracy (exrec > 6, parcomp = 0 or parcomp = 4 or parcomp = 5 but exrec != 8)

4-Full democracy (exrec = 8, parcomp = 5). See Goldstone et al. 2010 for full details. Numeric. Max = 5, min = 1, distinct = 6, mean = 2.425, sd = 1.478, NAs = 8089.

pitf_binary

A simplification of the pitf indicator of democracy described in Taylor and Ulfelder 2015. A country is a democracy (1) "if its chief executive is chosen in competitive elections (EXREC equal to 7 or 8) and political competition is not suppressed (PARCOMP equal to 0 or PARCOMP greater than 2)" Otherwise it is a non-democracy (0). Numeric. Max = 2, min = 1, distinct = 3, mean = 1.363, sd = 0.481, NAs = 8089.

polyarchy_contestation

1-9 index of contestation from the revised version of Coppedge and Reinicke 1991. Revised in 2003-2006. Includes a value for Western Sahara in 2000, which has been assigned code GWn 605; Western Sahara is not coded by any other dataset in this compilation, and is not considered an independent state by either Gleditsch and Ward or the Correlates of War project. Meaning of the scale is as follows:

9 Meaningful fair elections are held, there is full freedom for political organization and expression, and there is no preferential presentation of official views in the media.

8 Meaningful fair elections are held and there is full freedom for political organization and expression, but there is preferential presentation of official views in the media.

7 Meaningful fair elections are held and there is full freedom for political organization, but some public dissent is suppressed and there is preferential presentation of official views in the media.

6 Meaningful fair elections are held, but some independent political organizations are banned, some public dissent is suppressed, and there is preferential presentation of official views in the media.

5 Elections are marred by fraud or coercion, some independent political organizations are banned, some public dissent is suppressed, and there is preferential presentation of official views in the media.

4 Like score 5 except that there is less contestation in one or two of the following respects: no meaningful elections are held, only nonpolitical organizations are allowed to be independent, or alternatives to the official media are very limited.

3 No meaningful elections are held, only nonpolitical organizations are allowed to be independent, some public dissent is suppressed, and alternatives to the official media are very limited.

2 Like score 3 except that there is less contestation in one or two of the following respects: all organizations are banned or controlled by the government or official party, all public dissent is suppressed, or there is no public alternative to official information.

1 No meaningful elections are held, all organizations are banned or controlled by the government or official party, all public dissent is suppressed, and there is no public alternative to official information. Numeric. Max = 9, min = 1, distinct = 10, mean = 5.821, sd = 2.898, NAs = 23818.

prc

1-4 index of democracy from Gasiorowsk 1996. Available in updated form in Reich 2002. 1= Authoritarian, 2 = transitional, 3 = semidemocratic, 4 = democratic. Numeric. Max = 4, min = 1, distinct = 5, mean = 2.002, sd = 1.301, NAs = 13144.

prc_notrans

Same as prc but sets all 2 (transition) regimes to NA. Numeric. Max = 4, min = 1, distinct = 4, mean = 2.002, sd = 1.31, NAs = 13292.

svolik

Dichotomous indicator of democracy from Svolik 2012. 1 = authoritarian, 2 = democracy. This is extended for a few countries from the o_startdate variable in the original dataset. Numeric. Max = 2, min = 1, distinct = 3, mean = 1.439, sd = 0.496, NAs = 15403.

ulfelder

Dichotomous indicator of democracy from Ulfelder 2012. 0 = authoritarian, 1 = democracy. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.412, sd = 0.492, NAs = 16491.

utip_dichotomous

Calculated dichotomous index of democracy from data in the UTIP dataset of political regimes (Hsu 2008). 1 if the regime is a social democracy, conservative democracy, or one party democracy, 0 otherwise. The category of "one party democracy" is not well documented. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.515, sd = 0.5, NAs = 18132.

utip_dichotomous_strict

Stricter version of the calculated dichotomous index of democracy from data in the UTIP dataset of political regimes (Hsu 2008). 1 if the regime is a social democracy or a conservative democracy, 0 otherwise. This excludes "one party democracies" from the democracy category. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.477, sd = 0.5, NAs = 18132.

utip_trichotomous

Calculated trichotomous index of democracy from data in the UTIP dataset of political regimes (Hsu 2008). 2 if the regime is a social democracy or conservative democracy, 1 if the regime is a one party democracy, 0 otherwise. The category of "one party democracy" is not well documented. Numeric. Max = 2, min = 0, distinct = 4, mean = 0.992, sd = 0.981, NAs = 18132.

v2x_api

Additive polyarchy index from V-dem version 6.1 (Coppedge at al 2016). Higher values are more democratic. Numeric. Max = 0.983, min = 0.017, distinct = 9321, mean = 0.469, sd = 0.309, NAs = 8103.

v2x_delibdem

Deliberative democracy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.929, min = 0, distinct = 9885, mean = 0.213, sd = 0.267, NAs = 8211.

v2x_egaldem

Egalitarian democracy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.925, min = 0.007, distinct = 10213, mean = 0.246, sd = 0.246, NAs = 8103.

v2x_libdem

Liberal democracy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.928, min = 0.011, distinct = 10663, mean = 0.261, sd = 0.249, NAs = 8103.

v2x_mpi

Multiplicative polyarchy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.934, min = 0, distinct = 6115, mean = 0.176, sd = 0.279, NAs = 8103.

v2x_partipdem

Participatory democracy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.84, min = 0, distinct = 9950, mean = 0.2, sd = 0.205, NAs = 8112.

v2x_polyarchy

Continuous polyarchy index from V-dem version 6.1 (Coppedge et al 2016). Higher values are more democratic. Numeric. Max = 0.958, min = 0.008, distinct = 9321, mean = 0.323, sd = 0.282, NAs = 8103.

vanhanen_competition

Index of competition from Vanhanen 2012. From Vanahnen's documentation: "The smaller parties' share of the votes cast in parliamentary or presidential elections, or both, is used to indicate the degree of competition. It is calculated by subtracting the percentage of votes won by the largest party from 100. If the largest party gets, for example, 40 percent of the votes, the share of the smaller parties is 60 percent. If data on the distribution of votes are not available, the value of this variable is calculated on the basis of the distribution of seats in parliament. The distribution of seats is used also in cases in which it seems to indicate power relations more realistically than the distribution of votes." Its maximum value is 70. See the full documentation for Vanhanen's dataset for details. Numeric. Max = 70, min = 0, distinct = 695, mean = 25.217, sd = 25.171, NAs = 8440.

vanhanen_democratization

Index of democratization from Vanhanen 2012. Higher values are more democratic. Constructed multiplicatively from vanhanen_participation and vanhanen_competition. Numeric. Max = 49, min = 0, distinct = 446, mean = 8.434, sd = 11.676, NAs = 8440.

vanhanen_participation

Index of participation from Vanhanen 2012. From Vanahnen's documentation: "The percentage of the population which actually voted in the same elections is used to measure the degree of participation (= Participation). This percentage is calculated from the total population, not from the adult or enfranchized population." It is zero by construction in cases where no popular elections exist. May be modified by referenda. See the full documentation of Vanhanen's dataset for details. Numeric. Max = 71, min = 0, distinct = 761, mean = 21.099, sd = 21.877, NAs = 8440.

wahman_teorell_hadenius

Dichotomous measure of democracy from the Authoritarian Regimes Data Set, version 5.0 (Wahman, Teorell, and Hadenius 2013). Calculated from their regime1ny variable; non-democracy = all authoritarian regimes. Numeric. Max = 1, min = 0, distinct = 3, mean = 0.424, sd = 0.494, NAs = 17482.

References

Arat, Zehra F. 1991. Democracy and human rights in developing countries. Boulder: Lynne Rienner Publishers.

Bernhard, Michael, Timothy Nordstrom, and Christopher Reenock, "Economic Performance, Institutional Intermediation and Democratic Breakdown," Journal of Politics 63:3 (2001), pp. 775-803. Data and coding description available at http://users.clas.ufl.edu/bernhard/content/data/data.htm

Boix, Carles, Michael Miller, and Sebastian Rosato. 2012. A Complete Data Set of Political Regimes, 1800-2007. Comparative Political Studies 46 (12): 1523-1554. Original data available at https://sites.google.com/site/mkmtwo/democracy-v2.0.dta?attredirects=0.

Bollen, Kenneth A. 2001. "Cross-National Indicators of Liberal Democracy, 1950-1990." 2nd ICPSR version. Chapel Hill, NC: University of North Carolina, 1998. Ann Arbor, MI: Inter-university Consortium for Political and Social Research, 2001. Original data available at http://webapp.icpsr.umich.edu/cocoon/ICPSR-STUDY/02532.xml

Bowman, Kirk, Fabrice Lehoucq, and James Mahoney. 2005. Measuring Political Democracy: Case Expertise, Data Adequacy, and Central America. Comparative Political Studies 38 (8): 939-970. http://cps.sagepub.com/content/38/8/939. Data available at http://www.blmdemocracy.gatech.edu/.

Cheibub, Jose Antonio, Jennifer Gandhi, and James Raymond Vreeland. 2010. "Democracy and Dictatorship Revisited." Public Choice. 143(1):67-101. Original data available at https://sites.google.com/site/joseantoniocheibub/datasets/democracy-and-dictatorship-revisited.

Coppedge, Michael and Wolfgang H. Reinicke. 1991. Measuring Polyarchy. In On Measuring Democracy: Its Consequences and Concomitants, ed. Alex Inkeles. New Brunswuck, NJ: Transaction pp. 47-68.

Coppedge, Michael, John Gerring, Staffan I. Lindberg, Svend-Erik Skaaning, and Jan Teorell, with David Altman, Michael Bernhard, M. Steven Fish, Adam Glynn, Allen Hicken, Carl Henrik Knutsen, Kelly McMann, Pamela Paxton, Daniel Pemstein, Jeffrey Staton, Brigitte Zimmerman, Frida Andersson, Valeriya Mechkova, Farhad Miri. 2016. V-Dem Codebook v6.1. Varieties of Democracy (V-Dem) Project. Original data available at https://v-dem.net/en/data/.

Doorenspleet, Renske. 2000. Reassessing the Three Waves of Democratization. World Politics 52 (03): 384-406. DOI: 10.1017/S0043887100016580. http://dx.doi.org/10.1017/S0043887100016580.

Economist Intelligence Unit. 2012. Democracy Index 2012: Democracy at a Standstill.

Freedom House. 2015. "Freedom in the World." Original data available at http://www.freedomhouse.org.

Gasiorowski, Mark J. 1996. "An Overview of the Political Regime Change Dataset." Comparative Political Studies 29(4):469-483.

Geddes, Barbara, Joseph Wright, and Erica Frantz. 2014. Autocratic Breakdown and Regime Transitions: A New Data Set. Perspectives on Politics 12 (1): 313-331. Original data available at http://dictators.la.psu.edu/

Gleditsch, Kristian S. & Michael D. Ward. 1999. "Interstate System Membership: A Revised List of the Independent States since 1816." International Interactions 25: 393-413. The list can be found at http://privatewww.essex.ac.uk/~ksg/statelist.html

Goldstone, Jack, Robert Bates, David Epstein, Ted Gurr, Michael Lustik, Monty Marshall, Jay Ulfelder, and Mark Woodward. 2010. A Global Model for Forecasting Political Instability. American Journal of Political Science 54 (1): 190-208. DOI:10.1111/j.1540-5907.2009.00426.x

Hadenius, Axel & Jan Teorell. 2007. "Pathways from Authoritarianism", Journal of Democracy 18(1): 143-156.

Hadenius, Axel. 1992. Democracy and Development. Cambridge: Cambridge University Press.

Hsu, Sara "The Effect of Political Regimes on Inequality, 1963-2002," UTIP Working Paper No. 53 (2008), http://utip.gov.utexas.edu/papers/utip_53.pdf. Data available for download at http://utip.gov.utexas.edu/data/.

Kailitz, Steffen. 2013. Classifying political regimes revisited: legitimation and durability. Democratization 20 (1): 39-60. Original data available at http://dx.doi.org/10.1080/13510347.2013.738861.

Magaloni, Beatriz, Jonathan Chu, and Eric Min. 2013. Autocracies of the World, 1950-2012 (Version 1.0). Dataset, Stanford University. Original data and codebook available at http://cddrl.fsi.stanford.edu/research/autocracies_of_the_world_dataset/.

Mainwaring, Scott, Daniel Brinks, and Anibal Perez Linan. 2008. "Political Regimes in Latin America, 1900-2007." Original data available from http://kellogg.nd.edu/scottmainwaring/Political_Regimes.pdf.

Monty G. Marshall, Ted Robert Gurr, Keith Jaggers, 2014. POLITY IV PROJECT: Dataset Users' Manual. Center for Systemic Peace. Original dataset and codebook available at www.systemicpeace.org.

Moon, Bruce, Jennifer Harvey Birdsall, Sylvia Ceisluk, Lauren M. Garlett, Joshua J. Hermias, Elizabeth Mendenhall, Patrick D. Schmid, and Wai Hong Wong (2006) "Voting Counts: Participation in the Measurement of Democracy" Studies in Comparative International Development 42, 2 (Summer, 2006). The complete dataset is available here: http://www.lehigh.edu/~bm05/democracy/Obtain_data.htm.

Munck, Gerardo L. 2009. Measuring Democracy: A Bridge Between Scholarship and Politics. Baltimore: Johns Hopkins University Press.

Pemstein, Daniel, Stephen A. Meserve, and James Melton. 2013. "Replication data for: Democratic Compromise: A Latent Variable Analysis of Ten Measures of Regime Type." In: Harvard Dataverse. http://hdl.handle.net/1902.1/PMM.

Reich, G. 2002. Categorizing Political Regimes: New Data for Old Problems. Democratization 9 (4): 1-24. http://www.tandfonline.com/doi/pdf/10.1080/714000289.

Skaaning, Svend-Erik, John Gerring, and Henrikas Bartusevicius. 2015. A Lexical Index of Electoral Democracy. Comparative Political Studies 48 (12): 1491-1525.

Svolik, Milan. 2012. The Politics of Authoritarian Rule. Cambridge and New York: Cambridge University Press. Original data available from http://campuspress.yale.edu/svolik/the-politics-of-authoritarian-rule/

Taylor, Sean J. and Ulfelder, Jay, A Measurement Error Model of Dichotomous Democracy Status (May 20, 2015). Available at SSRN: http://ssrn.com/abstract=2726962 or http://dx.doi.org/10.2139/ssrn.2726962

Ulfelder, Jay. 2012. "Democracy/Autocracy Data Set." In: Harvard Dataverse. http://hdl.handle.net/1902.1/18836.

Vanhanen, Tatu. 2012. "FSD1289 Measures of Democracy 1810-2012." Original data available from http://www.fsd.uta.fi/english/data/catalogue/FSD1289/meF1289e.html.

Wahman, Michael, Jan Teorell, and Axel Hadenius. 2013. Authoritarian regime types revisited: updated data in comparative perspective. Contemporary Politics 19 (1): 19-34.

See also

Other democracy: all_gwf_extended_yearly, extended_uds, kailitz_yearly, lied, magaloni, polity_annual, vdem, wahman_teorell