Powell-Thyne coup data (V2015.09.30)

Count of coups by Powell and Thyne. Described in Powell and Thyne (2011); data is fully updated up to 2015. Original data available at http://www.uky.edu/~clthyn2/coup_data/powell_thyne_coups_final.txt

powell_thyne

Format

An object of class tbl_df (inherits from tbl, data.frame) with 474 rows and 18 columns.

Variable descriptions

country_name

Standardized country name. This is the same across all datasets in this package, so you can always join them by country_name and year. Character with 96 distinct values. Most common: Argentina (20), Bolivia (23), Sudan (14). NAs = 0.

GWn

Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999). Numeric. Max = 950, min = 40, distinct = 96, mean = 430.717, sd = 244.821, NAs = 0.

year

Year. Numeric. Max = 2015, min = 1950, distinct = 65, mean = 1976.749, sd = 15.714, NAs = 0.

date

Date of the coup. Date. Max = 2015-09-17, min = 1950-05-10, distinct = 468, NAs = 0.

coup

Coup type. 1 = unsuccessful, 2 = successful. Numeric. Max = 2, min = 1, distinct = 2, mean = 1.498, sd = 0.501, NAs = 0.

attempt_type

Character version of coup. Character with 2 distinct values. Most common: Successful (236), Unsuccessful (238). NAs = 0.

powell_country

Country name in the original dataset. Character with 97 distinct values. Most common: Argentina (20), Bolivia (23), Sudan (14). NAs = 0.

GWc

Gleditsch-Ward numeric country code. See Gleditsch and Ward (1999). Character with 96 distinct values. Most common: ARG (20), BOL (23), SUD (14). NAs = 0.

cown

Correlates of War numeric country code. Differs from GWn for a few country-years. See Gleditsch and Ward (1999). Numeric. Max = 950, min = 40, distinct = 96, mean = 430.717, sd = 244.821, NAs = 0.

region

Region. Character with 15 distinct values. Most common: South America (95), Western Africa (95), Western Asia (50). NAs = 0.

continent

Continent. Character with 5 distinct values. Most common: Africa (205), Americas (146), Asia (108). NAs = 0.

GW_startdate

Date at which the state entered the system of states according to Gleditsch and Ward, or NA if it has never been a member. Other.

GW_enddate

Date at which the state ceased to be a member of the system of states according to Gleditsch and Ward, or NA if it still exists. Other.

microstate

Indicator of whether the state is a microstate, according to Gleditsch's list of microstates. Logical. TRUE = 6, FALSE = 468, NAs = 0.

lat

Latitude. Numeric. Max = 61.524, min = -38.416, distinct = 96, mean = 7.889, sd = 18.345, NAs = 0.

lon

Longitude. Numeric. Max = 178.065, min = -90.231, distinct = 96, mean = 2.225, sd = 59.32, NAs = 0.

in_cow

Whether the country-year is in the Correlates of War system of states. Logical. TRUE = 462, FALSE = 12, NAs = 0.

in_system

Whether the country-year is in the Gleditsch-Ward system of states. See Gleditsch and Ward (1999). Logical. TRUE = 473, FALSE = 1, NAs = 0.

References

Powell, Jonathan M., and Clayton L. Thyne. 2011. "Global Instances of Coups from 1950 to 2010: A New Dataset." Journal of Peace Research 48 (2): 249-59.

See also

Other coups: PIPE

Other violence: fariss, navco_2